#33 Separate components for listings and cards
